@zadr 2019-01-24
don't mean to rush things by any means, but: ping? anyone have thoughts on this bit of code?