@jakob 2019-02-05
Hi!

Thanks for the pull request, this is pretty nice! I'm sorry that I didn't have any time to review this sooner, I've been pretty busy lately.

These are a lot of changes at once, so it'll take me a while to review in detail.

We probably can't use commit e12e156 -- the characterIndices are not the same as byte indices. It will only work for ASCII characters. For most other encodings (UTF-8, UTF-16) etc it will break.